+++ doc/source/whats_new.rst
@@ -111,6 +111,8 @@ Changelog
 
    - Add new somatosensory MEG dataset with nice time-frequency content by `Alex Gramfort`_
 
+   - Add InverseOperator class to display inverse operator info by `Mainak Jas`_
+
 
 BUG
 ~~~